+++ client/package.json
@@ -14,6 +14,7 @@
     "@react-navigation/bottom-tabs": "^6.5.4",
     "@react-navigation/native": "^6.1.3",
     "@react-navigation/native-stack": "^6.9.10",
+    "@react-query-firebase/firestore": "^1.0.0-dev.7",
     "axios": "^1.3.2",
     "expo": "~47.0.12",
     "expo-auth-session": "~3.8.0",
@@ -25,6 +26,7 @@
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
     "react-native-deck-swiper": "^2.0.12",
+    "react-native-fast-image": "^8.6.3",
     "react-native-gesture-handler": "~2.8.0",
     "react-native-reanimated": "~2.12.0",
     "react-native-safe-area-context": "4.4.1",
