+++ client/src/constants/QueryKey.ts
@@ -1,0 +1,3 @@
+export enum QueryKey {
+	LIKES = "likes",
+}
