+++ client/src/constants/index.ts
@@ -1,0 +1,1 @@
+export { QueryKey } from "./QueryKey";
