+++ client/src/types/RootStackParamList.ts
@@ -2,4 +2,5 @@ export type RootStackParamList = {
   Loading: undefined;
   Login: undefined;
   Home: undefined;
+  Settings: undefined;
 };
